Refactor test/utils.puy into package tests.utils
