+++ tests/test_actors2/test_power_meter.py
@@ -2,8 +2,8 @@ import argparse
 import asyncio
 import logging
 import typing
-from tests.fragment_runner import AsyncFragmentRunner
-from tests.fragment_runner import ProtocolFragment
+from tests.utils.fragment_runner import AsyncFragmentRunner
+from tests.utils.fragment_runner import ProtocolFragment
 from tests.utils import await_for
 
 import actors2
